tests/nested/cloud-init-{never-used,nocloud}-not-vuln: fix tests, use 2.45
Use snapd 2.45 as the baseline snapd version in the image that we build for the
refresh case, since this test variant is about old devices that were built 
before the fix was merged. 

Additionally, fix the environment variables which were subtly wrong (missing 
the _SNAPD_), and also replace the nested_is_core_16_system with the equivalent
os.query call.

Finally, fix the cloud-init journal check to actually fail if we don't see the
message after 60 seconds, and check for the right message (which is lower-case)


Signed-off-by: Ian Johnson <ian.johnson@canonical.com>