snap: add new `snap whoami --refresh-keys` to update ssh keys
This commit adds support for manually refresh the ssh keys of
the current user from Ubuntu SSO.

This is a first step to help with:
https://bugs.launchpad.net/snapd/+bug/1646559
so that at least after adding a key it's trivial to refresh it.
Users could enable a systemd user time to do this periodically
too.

Doing the real automatic syncing requires a bit more thinking

Doing the real automatic syncing requires a bit more thinking
(it seems to me) because then we would need to think about how
to detect if keys are managed by non-snapd and how to avoid
sending too many SSO requests etc.