packaging: fix build failure on bionic and simplify rules
For go-1.13 we need to ensure that we have a GOCACHE directory set.
If this is not set then the HOME directory is used. But in an sbuild
environment this is set to /sbuild-nonexistent so the build fails.

The reason we see this on 18.04 now is that we switched to the new
go-1.13 there but the dh-golang tooling is still assuming the 18.04
go-1.10.

This commit fixes this by setting one global GOCACHE.