{device,snap}state: skip kernel extraction in seeding
* {device,snap}state: skip kernel extraction in seeding

When seeding the device we currently extract the kernel. This is
unneeded because the kernel is already in place (without it the
system would not have been booted). Either ubuntu-image or "install"
mode extracted the kernel into the right place already.

This commit changes the code to skip kernel extraction when seeding.

* devicestate: make deciding of skipping for kernel extration the job of populateStateFromSeedImpl()

* snapstate: make SkipKernelExtraction part of snapstate.Flags

* snapstate: SetupSnapOpts->SetupSnapOptions (thanks to Miguel)