many: do not re-check snaps on disk during uc20 install
During the UC20 install the kernel and core20 snaps are put onto
the ubuntu-data partition by the MakeRunnableSystem() so that the
system can boot.

Then during seed both snaps are compared bit-by-bit with the
coresponding snaps in the seed partition by SetupSnap() and the
snapf.Install() call in there. This comparison is expensive, on
my test Pi3 it's almost 30s for the kernel and 22s for core20
(a really slow sd card though). And this check is also slightly
pointless because we used both snaps to boot. On encrypted systems
they are on the encrypted storage so there is little risk anyway.

So this commit adds a new "SkipOnDiskSnapValidation" (better name
suggestions welcome) that will simply do nothing if the target
snaps exist during seeding.

Mostly a PoC at this point to gather some real word data. Also
maybe not holistic enough in the approach.