gadget,secboot: add support to specify KDF options in the gadget
When deploying Ubuntu Core on specific boards the cost of benchmarking
what KDF parameters to use can be quite high. On one specific board
this benchmark adds ~20sec to the install. One possible solution
to avoid this delay is to measure the parameters in advance and then
put it into the gadget.yaml. This commit does that.

The downside of this approach is that it could lead to weak security
if people copy/paste gadget.yaml between boards that are differently
fast. So this branch also implements a mandatory "machine" line in
the gadget that much match something in /proc/cpuinfo. An alternative
approach would be to hash /proc/cpuinfo but it seems like there is
a bunch of things that are potentially unstable (bogoMIPS, fequency
scaling). It also does not look at /proc/meminfo:MemTotal but that
would be trivial to add.

With that the `gadget.yaml` on a Pi3 would look like this:
```
machine:
 kdf:
  memory-kib: 30000
  iterations: 4
  model: "Model:Raspberry Pi 3 Model B Plus Rev 1.3"

```

This is a draft, tests are missing etc, it's really there as a
discussion point.