tests/main/snapd-snap: Test apparmor interface policy generation
Duplicate the tests from interfaces-many-core-provided so we can test that
the snapd snap (with the vendored internal apparmor_parser etc) can
successfully compile the AppArmor policies which get generated by the
majority of the core provided interfaces. Note, this has some minor changes
compared to interfaces-many-core-provided as in this case the lxd snap is
also installed in this case.