snapstate: abort kernel refresh if no gadget update can be found
This commit detects if a kernel refresh is in progress that does
not contain a "update-gadget-assets" task and aborts this part
of the change if that happens. This works around the issue that
is triggered in https://bugs.launchpad.net/snapd/+bug/1940553

A manager test for this is also added.