snapstate: fix misleading `assumes` error message
When the `assumes` are not satisfied snapd will return an error
message that suggests to refresh `core` snap to fix the issue.

This is misleading and no longer valid. Hence this part is just
removed in this commit. It also unfies the message between
core and classic. Only on UC16 devices the message about refreshing
the core snap is still valid and those are in ESM mode.