many: shellcheck fixes
* many: shellcheck fixes

We got a new shellcheck in "edge" and it seems this broke master.
This commit adds the required fixes.

* tests: update `shellcheck disable=SC3037` for echo -e

We want to keep the `echo -e` - using printf is not cleaner and
also results in a shellcheck warning. We already had a shellcheck
disable that seems to have been changed with the new version of
shellcheck. So this commit updates it to the new SC3037 value.