sysconfig/cloudinit: don't run switch statement on error case
For the error case we should only concern ourselves with checking if the output
is "status: error", anything else should not be handled with the switch case,
as that could obscure the non-zero exit status with returning an otherwise
valid status. In other words, if cloud-init status returns "status: enabled"
but exits with non-zero exit code, we still want to return CloudInitErrored.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>