run-checks: fix incomplete gofmt
gofmt was only running on paths ending with "snapd/[A-Za-z0-9_]+$".
This was done to avoid running gofmt on root dirs and again in their
subdirs (since gofmt already does that) but would also exclude paths
that don't have a package in the dir directly below snapd, like snap/cmd.
snapd/cmd/snap (has a pkg) would be excluded and snapd/cmd isn't listed
by 'go list' (doesn't have a pkg) so it wasn't gofmt'ed.

This fix is simply runs gofmt on all dirs since the performance
impact was negligible (in my local tests).