tests: use "shellcheck" snap in `go` unit test
There is currently an inconsistency between when we run the
`tests/unit/go` shellcheck testing and the github action shellcheck.

The former will use whatever shellcheck is on the system but the
GH action will use the snap. That leads to the following issue:

With the new shellcheck we get an error SC3037 but the old shellcheck
from the distro calls that SC2039 - and we need to disable it.

This commit ensures that the go unit testis also using the snap
shellcheck.