tests: disable `SC2039` in `snapctl-hooks` configure hook
The new and the older shellcheck versions disagree about the
nature of the `echo -e` warning. New shellcheck warns with SC3037
and old with SC2039. The old shellcheck runs when in the
`tests/unit/go` spread test and the new in GH actions. Given that
this is the only instance where this is a problem this commit just
makes shellcheck ignore both the old and the new code for the issue.