many: move to yaml.v3
We wanted to move to `go-yaml/yaml.v3` for a while now. This commit
moves. It's mostly import changes but there are two actual changes
here:

The function `quotedIfNeeded` used `yaml.UnmarshalStrict()` which
is no longer available. However strict is not really needed here.

The `OrderedMap.UnmarshalYAML` has to be ported from `yaml.MapSlice`
to the new `yaml.Node`.

This is also useful for the netplan work because with `yaml.v3`
string yaml maps are unmarshalled as `map[string]interface{}`
which makes it compatible with json.