interfaces: add microstack-support interface
* Add microstack-support interface

Add an interface to enable MicroStack to work in a confined environment.

* Address review comments

* add rules necessary for LVM use

* Add rules required for the usage of iSCSI

OpenStack relies on the following components when working with LVM
volumes:

* tgtd - a daemon that exposes block devices via iSCSI;
* scsid - a control plane daemon for the iSCSI initiator data plane
implemented in the iscsi_tcp module;
* iscsi_tcp kernel module.

Working with the iSCSI kernel stack requires a more privileged access to
sysfs to that iscsi-adm and iscsid can do their job.

* Adjust the policy based on functional testing

After several tempest test runs, based on the kernel logs, it became
apparent that libvirt requires a wider rw access to the hierarchy under
/sys/fs/cgroup/*/machine.

* Fix test coverage for the microstack-support iface

* fix formatting in basedeclaration_test.go

* Fix comment typos/inconsistencies

* Replace tgtd rules with LIO ones

* fix a static check failure

* Address points raised in the review

* refactoring;
* added Delegate=true since libvirt manages its own cgroup subtree.

* Replace /dev/* rw with less privileged rules

The snapd team requested a less permissive rule to be used for block
device access. While MicroStack considers volume support experimental
due to https://bugs.launchpad.net/snapd/+bug/1892895, it seems
acceptable to limit the set of VG names to the ones prefixed with
"microstack-" to avoid blocking the whole review on this. Lifting this
naming restriction will be a consideration for the future versions of
this interface.

* Address review comments

* Fix indentation to avoid a static check failure

* TestKModConnectedPlug: expect the vfio-pci module

* Address security review comments

* Address review comments

* Remove the unused import