tests/main/snapd-snap: Workaround kernel memory leak issue on xenial
On the xenial 4.4 kernel, there appears to be a decent in-kernel memory
leak that occurs when reloading apparmor policies - over time this creates
a lot of memory pressure and results in apparmor_parser taking a large
amount of time to load policy whilst the kernel is busy coalescing free
pages etc. Work around this by increasing the overall kill-timeout for this
test and also by limiting the number of interfaces which will be tested in
any given run as is done on debian for the interfaces-many-core-provided
test already.