tests: ensure the `libvirt-daemon-system` pacakge is installed
The test for interfaces-libvirt needs the `libvirt-daemon-system`
that is currently not installed on our images for 20.04. This
broke with https://github.com/snapcore/snapd/pull/10675

The name of the group also changed from `libvirtd` to `libvirt`.

We could also add it to `pkgdb.sh` but given that this is really
only needed in this test it seems to be fine to just install it
here and remove it again.