sysconfig/cloudinit: add cloudDatasourcesInUseForDir
This function is the directory analogue of cloudDatasourcesInUse, and takes
into account the fact that a collection of config files coming from the same
dir will be processed by cloud-init "together" with lexically later files
overwriting settings of lexically earlier files.

This is necessary for full MAAS support with cloud-init on grade signed
systems.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>