client: turn the Client type into an interface
Create an interface for each source file which provides some public
methods on the Client type, which is turned into an interfaces composed
by all these bits.
This sets the ground for the introduction of a mock mechanism, through
which code depending on the Client type can be more easily tested by
mocking the client.