tests/lib/nested.sh: only wait for cloud-init if the VM is using cloud-init
If the VM is not using cloud-init, don't wait for it to be successful. This is
mainly needed to support the case in a future spread test where cloud-init is
not actually correctly setup for a nested VM in order to test the cloud-init
error case and the MAAS case.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>