snap packaging: Compile apparmor with a list of known AF names
When apparmor_parser is compiled, a list of known address family names is
created based on the installed system headers. Since the snapd snap is
built on Ubuntu 16.04 this list of known address families is quite a bit
out of date and hence the parser does not end up supporting newer address
families like AF_QIPCRTR. To work-around this, we include a pre-generated
af_names.h which has been generated on a Ubuntu 21.10 system so that it
includes the widest known list of families so they can be supported on any
system on which the snapd snap is installed.