tests: be more robust against a new day stepping in
Weird things can happen at midnight:

  https://pastebin.ubuntu.com/p/mmQskRXvhM/

The test could fail because it attempts to match the times by their date
component, without taking into account that if the two measurements
happen with the midnight in between, the date component will be
different.

Here we attempt to be a bit more robust, by comparing the given date not
only against the current date, but also against the date we had a few
minutes ago.