cmd/snap-confine: restore dlopen() to obtain udev_device_has_current_tag
Note that we could try to define udev_device_has_current_tag with a weak
attribute, which should in the normal case be the filled by ld.so when loading
snap-confined. However this was observed to work in practice only when the
binary itself is build with recent enough toolchain (eg. gcc & binutils on
Ubuntu 20.04). In case when the snapd snap needs to be built on 16.04, we have
observed that the weak symbol was indeed produced, but when resolving libraries
and symbols by ld.so on a recent host, the symbol would not be populated.

The patch revers to the original proposed behavior.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>