o/snapstate: only conflict with runnable and relevant tasks
We were signalling conflicts on tasks that affected the relevant snap but
couldn't run (e.g., failed), because we only checked if the Change had tasks
that could run. As a result, if two tasks sets in the same Change affected the
same snap and the first one failed, the second would always conflict because it
would find a relevant task that couldn't run but was in a change with tasks that could.