cmd/snap-confine: handle CURRENT_TAGS on systems that support it
* cmd/snap-confine: determine CURRENT_TAGS support, check current tags

Try to determine whether the libudev/libsystemd supports CURRENT_TAGS which has
been introduced in systemd v247:
https://github.com/systemd/systemd/blob/f6278558da0304ec6b646bb172ce4688c7f162a5/NEWS#L1037-L1119
This comes in pair with TAGS becoming sticky, what means that our checks no
longer behave correctly as the tags do not go away.

If the library supports current tags, use the dynamically collected symbol to
double check that the device is really currently tagged for a snap. The
implementation introduced in v247 correctly checks whether the running systemd
supports current tags and otherwise falls back to looking at TAGS property.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* tests/main/security-device-cgroups: tweak the tests to better handle sticky udev TAGS

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* cmd/snap-confine: let the linker fill the symbol

Instead of dlopen() and looking for the symbol ourselves, define a weak symbol
and let the dynamic linker do the work.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* cmd: snap-confine does not need to link with libdl anymore

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* cmd/snap-confine: tweak comment about current tags

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* tests/main/security-device-cgroup: tweak comments and messages

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* Revert "cmd: snap-confine does not need to link with libdl anymore"

This reverts commit 35d28c41d20d5f7d3451c4115041094db1f2ea4c.

* cmd/snap-confine: restore dlopen() to obtain udev_device_has_current_tag

Note that we could try to define udev_device_has_current_tag with a weak
attribute, which should in the normal case be the filled by ld.so when loading
snap-confined. However this was observed to work in practice only when the
binary itself is build with recent enough toolchain (eg. gcc & binutils on
Ubuntu 20.04). In case when the snapd snap needs to be built on 16.04, we have
observed that the weak symbol was indeed produced, but when resolving libraries
and symbols by ld.so on a recent host, the symbol would not be populated.

The patch revers to the original proposed behavior.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* tests/main/security-device-cgroups-serial-port: do not restart udev on 14.04

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>