overlord/devicestate: make settle wait longer in remodel tests
We occasionally see the TestUC20RemodelSetModelWithReboot fail when running the
test suite runs on Launchpad builder on armhf or s390, but not really on other
architectures. Use a longer timeout as the test seems to be working in general,
but it is suspected that the build host/vm is so slow that settle does not
converge in time.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>