spread: display information about current device cgroup in debug dump
This shoudl be regather impossible, but we see mknod failing with EPERM in
prepare of certain tests. As far as I can tell, the kernel should not be
triggering this with devmtmpfs even if the device the node we are creating for
does not exist. Perhaps there is a device cgroup set up for the current session.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>