gadget/install: add `encryptedDeviceWithSetupHook` device
This commit implements a encrypted device that is based on the
`device-setup` hook. The actual encryption setup is done via the
`device-setup` hook, the device itself is just a linear mapping
of the underlying partition with the first 1MB of data reserved
for future use (to add LUKS keyslot like meta-data).