osutil/disks: store DEVNAME and DEVPATH in *disk
This will be used later on when we need to take a Disk{} object and get the
kernel device path and kernel device node for that Disk.

This also means we don't need to query udev properties for the disk when we are
populating partitions, we already have the necessary information inside the
disk object itself.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>