osutil/disks: add Partition struct and Disks.Partitions()
This was also a long standing TODO to more fully represent a partition on a
disk, so here this is too.

We also refactor the FindMatchingPartitionUUIDWith{Fs,Part}Label methods to use
a more generic searching function which returns the full Partition not just the
PartitionUUID. The more generic methods will be added to the Disk interface
shortly.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>