tests/main/apparmor-batch-reload: Don't assume all features
When substituting our fake apparmor_parser, don't have it always return 0
when probing for features, otherwise during this test we may generate
apparmor profiles using this fake parser which appears to support all
features, but then these fail to compile when using the real
parser. Instead, let the real parser handle these invocations so that the
supported set of features remains the same as when using the real parser.

Also add support for the few new parser arguments that are used when snapd
is using the internal vendored apparmor_parser.