tests/main/snapd-snap: Use snap-mgmt --purge to revert snapd snap
Since we install the snapd snap we want to ensure it gets reverted after
this test runs so it is not used for other tests run by spread. However, on
most machines there is no snapd snap already installed so just blindly
calling revert fails - instead try using snap-mgmt --purge to revert this
for subsequent tests.